Complete synchronous abort method modeled after the existing ones.
Because our USB stack wants the aborted xfer to be removed from the
pipe during abort(), we have to sleep in the abort function.

Regarding the xHCI process, when a TD is being aborted, we simply stop
the endpoint and then move the dequeue pointer past its last TRB.  This
is fairly simple for the moment since only one xfer can be pending on a
given pipe.