Use rt_ifa_add(9) and rt_ifa_del(9) to configure MPLS labels.
Since the removal of the link-layer ifa from "struct ifnet" it was
impossible to configure MPLS routes on mpe(4).  Because just like
enc(4), mpe(4) was abusing the link-layer ifa to attach its route
entries.

So explicitly pass a "glue" ifa to the routing layer.

From Rafael Zalamena who discovered this the hardway, thanks!

ok claudio@