move to a per rx ring timeout for refilling empty rings.
this lets me get rid of the locking around the refilling of the rx ring.

the timeout only runs refill if the rx ring is empty. we know rxeof
wont try and refill it in that situation because there's no packets
on the ring so we wont get interrupts for it. therefore we dont
need to lock between the timeout and rxeof cos they cant run at the
same time.